Rename files to match current convention.
