+++ lib/idGenerator.js
@@ -7,7 +7,7 @@ const async = require('async');
 const mongo = require('mongodb');
 const util = require('util');
 const bedrock = require('bedrock');
-const database = require('./database');
+const database = require('./index');
 
 const api = {};
 module.exports = api;
