+++ lib/index.js
