+++ lib/migration.js
@@ -7,7 +7,7 @@ const _ = require('lodash');
 const async = require('async');
 const bedrock = require('bedrock');
 const config = bedrock.config;
-const database = require('./database');
+const database = require('./index');
 const progress = require('progress');
 
 const BedrockError = bedrock.util.BedrockError;
