+++ package.json
@@ -2,10 +2,7 @@
   "name": "bedrock-mongodb",
   "version": "3.2.4-0",
   "description": "Bedrock mongodb module",
-  "main": "./lib/database",
-  "scripts": {
-    "test": "node test.js"
-  },
+  "main": "lib",
   "repository": {
     "type": "git",
     "url": "https://github.com/digitalbazaar/bedrock-mongodb"
